+++ viadot/flows/supermetrics_to_adls.py
@@ -1,15 +1,20 @@
 import os
+import pendulum
+import pyarrow
+import json
 from typing import Any, Dict, List, Union
 
-import pandas as pd
-import pendulum
-from prefect import Flow, Task, apply_map, task
-from prefect.storage import Git, GitHub, Local
-from prefect.tasks.control_flow import case
-from prefect.tasks.secrets import PrefectSecret
-from prefect.utilities import logging
+from visions.functional import infer_type
+from visions.typesets.complete_set import CompleteSet
 
-from ..task_utils import add_ingestion_metadata_task
+import pandas as pd
+from prefect import Flow, Task, apply_map, task
+from prefect.storage import Git, GitHub
+from prefect.tasks.control_flow import case
+from prefect.utilities import logging
+from prefect.backend import set_key_value, get_key_value
+
+from ..task_utils_v2 import METADATA_COLUMNS, add_ingestion_metadata_task
 from ..tasks import (
     AzureDataLakeUpload,
     DownloadGitHubFile,
@@ -22,7 +27,13 @@ logger = logging.get_logger(__name__)
 supermetrics_to_df_task = SupermetricsToDF()
 download_github_file_task = DownloadGitHubFile()
 validation_task = RunGreatExpectationsValidation()
-csv_to_adls_task = AzureDataLakeUpload()
+parquet_to_adls_task = AzureDataLakeUpload()
+json_to_adls_task = AzureDataLakeUpload()
+
+
+@task
+def get_data_types(df: pd.DataFrame) -> dict:
+    df.dtypes.to_dict()
 
 
 @task
@@ -31,16 +42,38 @@ def union_dfs_task(dfs: List[pd.DataFrame]):
 
 
 @task
-def df_to_csv_task(df, path: str, sep: str = "\t"):
-    df.to_csv(path, sep=sep, index=False)
+def df_get_data_types_task(df):
+    typeset = CompleteSet()
+    dtypes = infer_type(df, typeset)
+    dtypes_dict = {k: str(v) for k, v in dtypes.items()}
+    return dtypes_dict
+
+
+@task
+def dtypes_to_json_task(dtypes_dict, local_json_path: str):
+    with open(local_json_path, "w") as fp:
+        json.dump(dtypes_dict, fp)
+
+
+@task
+def df_to_parquet_task(df, path: str, if_exists: str = "replace"):
+    if if_exists == "append":
+        if os.path.isfile(path):
+            parquet_df = pd.read_parquet(path)
+            out_df = pd.concat([parquet_df, df])
+        else:
+            out_df = df
+    elif if_exists == "replace":
+        out_df = df
+    out_df.to_parquet(path, index=False)
 
 
 @task
 def is_stored_locally(f: Flow):
-    return f.storage is None or isinstance(f.storage, Local)
+    return f.storage is None
 
 
-class SupermetricsToAdls(Flow):
+class SupermetricsToADLS(Flow):
     def __init__(
         self,
         name: str,
@@ -58,7 +91,6 @@ class SupermetricsToAdls(Flow):
         expectation_suite_name: str = "failure",
         local_file_path: str = None,
         adls_dir_path: str = None,
-        sep: str = "\t",
         overwrite_adls: bool = True,
         if_empty: str = "warn",
         if_exists: str = "replace",
@@ -104,14 +136,6 @@ class SupermetricsToAdls(Flow):
             tags (List[str], optional): Flow tags to use, eg. to control flow concurrency. Defaults to ["extract"].
             vault_name (str, optional): The name of the vault from which to obtain the secrets. Defaults to None.
         """
-
-        if not ds_user:
-            try:
-                ds_user = PrefectSecret("SUPERMETRICS_DEFAULT_USER").run()
-            except ValueError as e:
-                msg = "Neither 'ds_user' parameter nor 'SUPERMETRICS_DEFAULT_USER' secret were not specified"
-                raise ValueError(msg) from e
-
         # SupermetricsToDF
         self.ds_id = ds_id
         self.ds_accounts = ds_accounts
@@ -129,11 +153,15 @@ class SupermetricsToAdls(Flow):
         # RunGreatExpectationsValidation
 
         # AzureDataLakeUpload
-        self.local_file_path = local_file_path or self.slugify(name) + ".csv"
-        self.adls_dir_path = os.path.join(
-            adls_dir_path, str(pendulum.now("utc")) + ".csv"
+
+        self.local_file_path = local_file_path or self.slugify(name) + ".parquet"
+        self.local_json_path = self.slugify(name) + ".json"
+        self.now = str(pendulum.now("utc"))
+        self.adls_dir_path = adls_dir_path
+        self.adls_file_path = os.path.join(adls_dir_path, self.now + ".parquet")
+        self.adls_schema_file_dir_file = os.path.join(
+            adls_dir_path, "schema", self.now + ".json"
         )
-        self.sep = sep
         self.overwrite_adls = overwrite_adls
         self.if_empty = if_empty
         self.adls_sp_credentials_secret = adls_sp_credentials_secret
@@ -156,6 +184,7 @@ class SupermetricsToAdls(Flow):
         )
         self.storage_repo = self._get_expectation_suite_repo()
 
+        # self.dtypes.update(METADATA_COLUMNS)
         self.gen_flow()
 
     @staticmethod
@@ -163,7 +192,7 @@ class SupermetricsToAdls(Flow):
         return name.replace(" ", "_").lower()
 
     def _get_expectation_suite_repo(self):
-        if self.storage is None or isinstance(self.storage, Local):
+        if self.storage is None:
             return None
         return self.storage.repo
 
@@ -171,23 +200,23 @@ class SupermetricsToAdls(Flow):
 
         if self.storage is None:
             return os.path.join(os.getcwd(), self.expectation_suite_file_name)
-        elif isinstance(self.storage, GitHub):
-            path = self.storage.path
-        elif isinstance(self.storage, Git):
-            # assuming this is DevOps
-            path = self.storage.flow_path
-        elif isinstance(self.storage, Local):
-            path = self.storage.directory
-        else:
-            try:
-                path = self.storage.path
-            except AttributeError:
-                raise NotImplemented("Unsupported storage type.")
 
-        flow_dir_path = path[: path.rfind("/")]
-        return os.path.join(
-            flow_dir_path, "expectations", self.expectation_suite_file_name
-        )
+        else:
+            if isinstance(self.storage, GitHub):
+                path = self.storage.path
+            elif isinstance(self.storage, Git):
+                # assuming this is DevOps
+                path = self.storage.flow_path
+            else:
+                try:
+                    path = self.storage.path
+                except AttributeError:
+                    raise NotImplemented("Unsupported storage type.")
+
+            flow_dir_path = path[: path.rfind("/")]
+            return os.path.join(
+                flow_dir_path, "expectations", self.expectation_suite_file_name
+            )
 
     def gen_supermetrics_task(
         self, ds_accounts: Union[str, List[str]], flow: Flow = None
@@ -228,38 +257,48 @@ class SupermetricsToAdls(Flow):
         )
 
         local = is_stored_locally.bind(self, flow=self)
-        with case(local, True):
-            pass_the_other_case = True
-        if not pass_the_other_case:
-            with case(local, False):
-                download_expectations = download_github_file_task.bind(
-                    repo=self.storage_repo,
-                    from_path=self.expectation_suite_path,
-                    to_path=self.expectation_suite_local_path,
-                    flow=self,
-                )
-                validation.set_upstream(download_expectations, flow=self)
+        with case(local, False):
+            download_expectations = download_github_file_task.bind(
+                repo=self.storage_repo,
+                from_path=self.expectation_suite_path,
+                to_path=self.expectation_suite_local_path,
+                flow=self,
+            )
+            validation.set_upstream(download_expectations, flow=self)
 
-        df_to_csv = df_to_csv_task.bind(
-            df=df,
+        df_with_metadata = add_ingestion_metadata_task(df, flow=self)
+
+        # get_data_types = get_data_types(df_with_metadata)
+
+        df_to_parquet = df_to_parquet_task.bind(
+            df=df_with_metadata,
             path=self.local_file_path,
-            sep=self.sep,
+            if_exists=self.if_exists,
             flow=self,
         )
 
-        add_ingestion_metadata = add_ingestion_metadata_task.bind(
-            path=self.local_file_path, sep=self.sep, flow=self
-        )
-
-        csv_to_adls_task.bind(
+        parquet_to_adls_task.bind(
             from_path=self.local_file_path,
-            to_path=self.adls_dir_path,
+            to_path=self.adls_file_path,
             overwrite=self.overwrite_adls,
             sp_credentials_secret=self.adls_sp_credentials_secret,
             vault_name=self.vault_name,
             flow=self,
         )
-
-        df_to_csv.set_upstream(validation, flow=self)
-        add_ingestion_metadata.set_upstream(df_to_csv, flow=self)
-        csv_to_adls_task.set_upstream(add_ingestion_metadata_task, flow=self)
+        dtypes_dict = df_get_data_types_task.bind(df_with_metadata, flow=self)
+        dtypes_to_json_task.bind(
+            dtypes_dict=dtypes_dict, local_json_path=self.local_json_path, flow=self
+        )
+        json_to_adls_task.bind(
+            from_path=self.local_json_path,
+            to_path=self.adls_schema_file_dir_file,
+            overwrite=self.overwrite_adls,
+            sp_credentials_secret=self.adls_sp_credentials_secret,
+            vault_name=self.vault_name,
+            flow=self,
+        )
+        df_with_metadata.set_upstream(validation, flow=self)
+        df_to_parquet.set_upstream(df_with_metadata, flow=self)
+        parquet_to_adls_task.set_upstream(df_to_parquet, flow=self)
+        json_to_adls_task.set_upstream(dtypes_to_json_task, flow=self)
+        set_key_value(key=self.adls_dir_path, value=self.adls_file_path)
